+++ bin/execute.js
@@ -1,18 +1,20 @@
-#!/usr/bin/env node
-const inquirer = require('inquirer');
-const semver = require('semver');
-const package = require('../package.json');
+#!/usr/bin/env node --no-warnings
+import semver from 'semver';
+import packageJSON from '../package.json' assert {
+	type: 'json'
+};
 
-const engineVersion = package.engines.node;
+const engineVersion = packageJSON.engines.node;
 
 if (!semver.satisfies(process.version, engineVersion)) {
   console.log(`Required node version ${engineVersion} not satisfied with current version ${process.version}.`);
   process.exit(1);
 }
 
-const execute = require('../out/execute').default;
-const program = require('commander').program;
-const version = package.version;
+import execute from '../out/execute.js';
+import { program } from 'commander';
+
+const version = packageJSON.version;
 let options = {};
 
 program.version(version);
@@ -36,19 +38,14 @@ program.parse(process.argv);
 	options = Object.assign(options, program.opts());
 
 	if (options.interactive) {
-		options.params = await inquirer
-			.prompt({
-				name: 'default',
-				message: 'Params:',
-				type: 'input',
-				loop: false
-			})
-			.then((inputMap) => {
-				return inputMap.default.split(' ');
-			})
-			.catch((err) => {
-				throw err;
-			});
+		const inquirer = require('inquirer');
+		const interactiveParams = await inquirer.prompt({
+			name: 'default',
+			message: 'Params:',
+			type: 'input'
+		});
+
+		options.params = interactiveParams.default.split(' ');
 	}
 
 	const success = await execute(options.filepath, {