+++ src/build.ts
@@ -7,7 +7,7 @@ import {
 import mkdirp from 'mkdirp';
 import path from 'path';
 
-import EnvMapper from './build/env-mapper';
+import EnvMapper from './build/env-mapper.js';
 
 function createContentHeader(): string {
   return ['s=get_shell', 'c=s.host_computer', 'h=home_dir', 'p=@push'].join(
@@ -257,7 +257,7 @@ export default async function build(
 
     const getRelativePath = (filePath: string) => {
       const pathSegments = filePath.split(path.sep);
-      const filtered = [];
+      const filtered: string[] = [];
 
       for (const segment of targetRootSegments) {
         const current = pathSegments.shift();