+++ src/repl/debugger.ts
@@ -1,0 +1,13 @@
+import { Debugger, DefaultType, OperationContext } from 'greybel-interpreter';
+
+export default class GrebyelPseudoDebugger extends Debugger {
+  debug() {
+    return DefaultType.Void;
+  }
+
+  getBreakpoint(_operationContext: OperationContext): boolean {
+    return false;
+  }
+
+  interact(_operationContext: OperationContext) {}
+}