+++ src/web/extension/grammar/language.ts
@@ -1,5 +1,5 @@
-import { allTypes, getDefinitions } from 'greyscript-meta/dist/meta';
-import Monaco from 'monaco-editor/esm/vs/editor/editor.api';
+import { allTypes, getDefinitions } from 'greyscript-meta/dist/meta.js';
+import Monaco from 'monaco-editor/esm/vs/editor/editor.api.js';
 
 const language: Monaco.languages.IMonarchLanguage = {
   // setup via https://microsoft.github.io/monaco-editor/monarch.html