+++ CHANGELOG.md
@@ -514,3 +514,9 @@ All notable changes to this project will be documented in this file.
 
 - returning proper library type names when using typeof
 - time now returns elapsed time in seconds
+
+## [1.7.7] - 30.04.2023
+
+- replace newline in print with actual newline
+- use editor inquirer component for repl
+- refactor output handler enabling user_input not forcing newline